backend/requirements.txt
alembic==1.13.2
asttokens==2.4.1
blinker==1.8.2
branca==0.7.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
executing==2.0.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
flask-shell-ipython==0.5.1
Flask-SQLAlchemy==3.1.1
folium==0.17.0
greenlet==3.0.3
idna==3.7
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
numpy==2.0.1
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyproj==3.6.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.32
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
xyzservices==2024.6.0
